support custom code

+++ packages/generators/cypress/src/code.ts
@@ -72,14 +72,6 @@ export default class CypressCodeGenerator {
       fs.mkdirSync(this.localTestFolder);
     }
 
-    if (fs.existsSync(this.localSupportFolder)) {
-      fs.rmdirSync(this.localSupportFolder, { recursive: true });
-    }
-
-    if (fs.existsSync(this.localTestFolder)) {
-      fs.rmdirSync(this.localTestFolder, { recursive: true });
-    }
-
     if (
       !fs.existsSync(this.localSupportFolder) &&
       fs.existsSync(this.outputPath)
@@ -411,43 +403,72 @@ cy.get(${`${constants.LOCATOR_KEY_WORD}.${this.sanitizeKey(
     const e2eFile = this.workspacePath
       ? `${this.localSupportFolder}/e2e.js`
       : `${this.localSupportFolder}/e2e.ts`;
-    fs.appendFileSync(e2eFile, `import 'cypress-network-idle';`);
-    fs.appendFileSync(
-      e2eFile,
+
+    const data = fs.existsSync(e2eFile)
+      ? fs.readFileSync(e2eFile, 'utf8')
+      : undefined;
+
+    if (data && data.includes(constants.MARVIN_GENERATED_COMMENT)) {
+      fs.unlinkSync(e2eFile);
+    }
+
+    if (!fs.existsSync(e2eFile)) {
+      fs.appendFileSync(e2eFile, `${constants.MARVIN_GENERATED_COMMENT} \r\n`);
+      fs.appendFileSync(e2eFile, `import 'cypress-network-idle';`);
+      fs.appendFileSync(
+        e2eFile,
+        `
+      Cypress.on('uncaught:exception', (err, runnable) => {
+        return false;
+      });
+      Cypress.on('load', (err, runnable) => {
+        return false;
+      });
       `
-    Cypress.on('uncaught:exception', (err, runnable) => {
-      return false;
-    });
-    Cypress.on('load', (err, runnable) => {
-      return false;
-    });
-    `
-    );
+      );
+      for (const command of commands) {
+        const { file } = command;
+        fs.appendFileSync(
+          e2eFile,
+          `import './commands/${file}';
+          `
+        );
+      }
+      this.formatFile(e2eFile);
+    }
     for (const command of commands) {
       const { file, methods } = command;
       const commandFile = `${this.localSupportFolder}/commands/${file}`;
-
-      fs.appendFileSync(
-        e2eFile,
-        `import './commands/${file}';
-      `
-      );
       let importLocation: string = this.getRelativePath(
         `${this.localSupportFolder}/commands`,
         `${this.localSupportFolder}/app.po.js`
       );
-      fs.writeFileSync(
-        commandFile,
-        `
-  import * as locators from '${importLocation}';
-        `
-      );
-      for (const method of methods) {
-        await this.writeMethod(commandFile, method);
+
+      const commandData = fs.existsSync(commandFile)
+        ? fs.readFileSync(commandFile, 'utf8')
+        : undefined;
+
+      if (
+        commandData &&
+        commandData.includes(constants.MARVIN_GENERATED_COMMENT)
+      ) {
+        fs.unlinkSync(commandFile);
       }
-      this.formatFile(commandFile);
+
+      if (!fs.existsSync(commandFile)) {
+        fs.writeFileSync(
+          commandFile,
+          `
+          ${constants.MARVIN_GENERATED_COMMENT} \n\r
+    import * as locators from '${importLocation}';
+          `
+        );
+        for (const method of methods) {
+          await this.writeMethod(commandFile, method);
+        }
+        this.formatFile(commandFile);
+      }
     }
-    this.formatFile(e2eFile);
   }
 
   private replaceKeyWord(selector: string): string {
@@ -515,11 +536,24 @@ cy.get(${`${constants.LOCATOR_KEY_WORD}.${this.sanitizeKey(
       ? `${this.localSupportFolder}/app.po.js`
       : `${this.localSupportFolder}/app.po.ts`;
     let fileContent: string = '';
-    for (const selector of selectors) {
-      fileContent = `
-export const ${this.sanitizeKey(selector.key)} = '${selector.value}';
-      `;
-      fs.appendFileSync(selectorFile, fileContent);
+    const data = fs.existsSync(selectorFile)
+      ? fs.readFileSync(selectorFile, 'utf8')
+      : undefined;
+    if (data && data.includes(constants.MARVIN_GENERATED_COMMENT)) {
+      fs.unlinkSync(selectorFile);
+    }
+
+    if (!fs.existsSync(selectorFile)) {
+      fs.appendFileSync(
+        selectorFile,
+        `${constants.MARVIN_GENERATED_COMMENT} \n\r`
+      );
+      for (const selector of selectors) {
+        fileContent = `
+  export const ${this.sanitizeKey(selector.key)} = '${selector.value}';
+        `;
+        fs.appendFileSync(selectorFile, fileContent);
+      }
     }
   }
 
@@ -614,6 +648,13 @@ export const ${this.sanitizeKey(selector.key)} = '${selector.value}';
       for (const spec of specs) {
         const { file, beforeAll, tests } = spec;
         const specFile = `${groupFolder}/${file}`;
+        const data = fs.existsSync(specFile)
+          ? fs.readFileSync(specFile, 'utf8')
+          : undefined;
+
+        if (data && data.includes(constants.MARVIN_GENERATED_COMMENT)) {
+          fs.unlinkSync(specFile);
+        }
         let startDescribeCommand = `
         
         describe('${file}', () => {
@@ -651,14 +692,17 @@ export const ${this.sanitizeKey(selector.key)} = '${selector.value}';
         });`;
 
         const finalOutput =
+          `${constants.MARVIN_GENERATED_COMMENT}` +
           startDescribeCommand +
           beforeAllContent +
           this.getTestsBody(tests).join('\r\n') +
           endDescribeCommand;
-        fs.writeFileSync(
-          specFile,
-          prettier.format(finalOutput, { parser: 'typescript' })
-        );
+        if (!fs.existsSync(specFile)) {
+          fs.writeFileSync(
+            specFile,
+            prettier.format(finalOutput, { parser: 'typescript' })
+          );
+        }
       }
     }
   }
